+++ acfd.php
@@ -1,23 +1,24 @@
 <?php
 /*
-Plugin Name: Advanced Custom Fields Easy Development
-Plugin URI: http://www.mirkoferraro.it
+Plugin Name: Advanced Custom Fields Development
+Plugin URI: https://github.com/mirkoferraro/advanced-custom-fields-dev
 Description: Makes the ACF registration via PHP easier
-Version: 1.0.2
+Version: 1.0.3
 Author: Mirko Ferraro
 Author URI: http://www.mirkoferraro.it
 Copyright: Mirko Ferraro
 */
 
-if( ! defined( 'ABSPATH' ) ) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-class ACFED {
+class ACFD {
 	
 	static private $menu_order = 1;
-	static private $defaults = array();
-	static private $groups = array();
+	static private $defaults   = array();
+	static private $groups     = array();
+	static private $customfile = null;
 
 	static function init() {
 		add_action( 'plugins_loaded', array( get_called_class(), 'check_dependencies' ) );
@@ -32,12 +33,15 @@ class ACFED {
 		include( 'class/custom_field.php' );
 		include( 'class/custom_field_container.php' );
 		include( 'class/custom_group.php' );
+		include( 'class/custom_module.php' );
 		include( 'class/widget.php' );
+		
+		include( 'includes/hide_on_screen_fix.php' );
 
 		add_action('acf/init', array( get_called_class(), 'registerGroups' ) );
 	}
 
-	static function is_active() {
+	static function isActive() {
 		return class_exists( 'acf' );
 	}
 
@@ -49,9 +53,28 @@ class ACFED {
 	    <?php
 	}
 
+	static function setCustomFile( $path ) {
+		self::$customfile = $path;
+	}
+
+	static function loadCustomFile() {
+		if ( self::customFileExists() ) {
+			include_once( self::$customfile );
+			return true;
+		}
+
+		return false;
+	}
+
+	static function customFileExists() {
+		return file_exists( self::$customfile );
+	}
+
 	static function getContainerFieldName( $type ) {
 		switch ( $type ) {
 			case 'repeater': return 'sub_fields';
+			case 'flexible_content': return 'layouts';
+			case 'flexible_item': return 'sub_fields';
 			default: return 'fields';
 		}
 	}
@@ -64,6 +87,22 @@ class ACFED {
 		self::$defaults[ $type ] = $data;
 	}
 
+	static function field( $name, $label, $type ) {
+		return new CustomField( $name, $label, $type );
+	}
+
+	static function container( $name, $label, $type, $field_name = 'fields' ) {
+		return new CustomFieldContainer( $name, $label, $type, $field_name );
+	}
+
+	static function group( $name, $location = 'options_page == acf-options' ) {
+		return new CustomGroup( $name, $location );
+	}
+
+	static function module() {
+		return new CustomModule();
+	}
+
 	static function addGroup( $group ) {
 		self::$groups[] = $group;
 	}
@@ -74,9 +113,12 @@ class ACFED {
 			throw new Exception('Advanced Custom Field not loaded');
 		}
 
+		$fields_code = '';
+		$options_code = '';
+
 		foreach ( self::$groups as $group ) {
 
-			if ( ! count( $group->fields() ) ) {
+			if ( ! count( $group->getFields() ) ) {
 				continue;
 			}
 			
@@ -88,21 +130,31 @@ class ACFED {
 
 			acf_add_local_field_group( $group_data );
 
+			if ( self::$customfile ) {
+				$fields_code .= ' acf_add_local_field_group(' . var_export( $group_data, true ) . ');';
+			}
+
 			// Create ACF Options page if used
 			foreach ( $group_data['location'] as $group ) {
 				foreach ($group as $rule) {
 					if ( $rule['param'] == 'options_page' && $rule['operator'] == '==' ) {
 						if ( strpos( $rule['value'], 'acf-options-' ) === 0 ) {
 							acf_add_options_sub_page( substr( $rule['value'], 12 ) );
+							$file_code .= ' acf_add_options_sub_page("' . substr( $rule['value'], 12 ) . '");';
 						} else {
 							acf_add_options_page();
+							$file_code .= ' acf_add_options_page();';
 						}
 					}
 				}
 			}
 		}
 
+		if ( self::$customfile ) {
+			$phpcode = '<?php ' . $options_code . $fields_code;
+			file_put_contents( self::$customfile, $phpcode );
+		}
 	}
 }
 
-ACFED::init();
+ACFD::init();