+++ class/custom_field_container.php
@@ -32,9 +32,7 @@ class CustomFieldContainer extends CustomField {
 		$fields_data = array();
 
 		foreach ( $this->fields as $field ) {
-			if ( count( $field->fields ) ) {
-				$fields_data[] = $field->get();
-			}
+			$fields_data[] = $field->get();
 		}
 
 		return $fields_data;
@@ -58,12 +56,28 @@ class CustomFieldContainer extends CustomField {
 
 	function addContainer( $name, $label, $type, $field_name = null ) {
 		if ( $field_name == null ) {
-			$field_name = ACFED::getContainerFieldName( $type );
+			$field_name = ACFD::getContainerFieldName( $type );
 		}
 
 		$field = new CustomFieldContainer( $name, $label, $type, $field_name );
 		$this->fields[] = $field;
 		return $field;
+	}
+
+	function addModule( $module, $prefix = null ) {
+		if ( $module == null ) {
+			return $this;
+		}
+
+		if ( $prefix == null ) {
+			$prefix = $this->get( 'name' ) . '_';
+		}
+		
+		foreach ( $module->fields as $field ) {
+			$this->addField( $field->clone( $prefix . $field->get( 'name' ) ) );
+		}
+
+		return $this;
 	}
 	
 }