+++ class/custom_module.php
@@ -1,0 +1,22 @@
+<?php
+
+class CustomModule extends CustomFieldContainer {
+
+	function __construct() {
+	}
+
+	function get( $key = null ) {
+		if ( $key != null ) {
+			if ( $key == $this->field_name ) {
+				return $fields;
+			}
+
+			return parent::get( $key );
+		}
+
+		return $this->getFieldsData();
+	}
+
+}
+
+