+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.0.3
+	Changed plugin name to ACFD
+	Added CustomModule class that allow you to re-use custom fields
+	Added some useful tips in the readme
+	Fix ACF issue on multiple hide_on_screen function
+
 # 1.0.2
 	Now you can pass an instance of CustomField to the addField function